PDJB-1279: Co-locate org landlord joint flow tests with their counterparts
