+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/CancelJointLandlordInvitationJourneyTests.kt
@@ -60,3 +60,50 @@ class CancelJointLandlordInvitationJourneyTests :
         assertThat(propertyDetailsPage.pendingInvitationsDetails).hasCount(0)
     }
 }
+
+@WithOrgLandlordProfile
+class JointLandlordInvitationOrgLandlordJourneyTests : IntegrationTestWithMutableData("data-local.sql") {
+    private val propertyOwnershipId = 48L
+    private val invitedEmail = "jl.pending.org@example.com"
+
+    @Test
+    fun `An org landlord can cancel a pending invitation on their property`(page: Page) {
+        val detailsPage = navigator.goToPropertyDetailsLandlordView(propertyOwnershipId)
+        detailsPage.tabs.goToLandlordDetails()
+        detailsPage.pendingInvitationsSummary.click()
+        detailsPage.cancelInvitationLink.clickAndWait()
+
+        val areYouSurePage = assertPageIs(page, AreYouSurePageCancelJointLandlordInvitation::class)
+        assertThat(page.locator("main")).containsText(invitedEmail)
+        areYouSurePage.submitWantsToProceed()
+
+        val confirmationPage = assertPageIs(page, ConfirmationPageCancelJointLandlordInvitation::class)
+        assertThat(confirmationPage.confirmationBanner).containsText(invitedEmail)
+        confirmationPage.goBackToPropertyRecordLink.clickAndWait()
+
+        val propertyDetailsPage =
+            assertPageIs(
+                page,
+                PropertyDetailsPageLandlordView::class,
+                mapOf("propertyOwnershipId" to propertyOwnershipId.toString()),
+            )
+        propertyDetailsPage.tabs.goToLandlordDetails()
+        assertThat(propertyDetailsPage.pendingInvitationsDetails).hasCount(0)
+    }
+
+    @Test
+    fun `An org landlord can resend a pending invitation on their property`(page: Page) {
+        val detailsPage = navigator.goToPropertyDetailsLandlordView(propertyOwnershipId)
+        detailsPage.tabs.goToLandlordDetails()
+        detailsPage.pendingInvitationsSummary.click()
+        detailsPage.resendInvitationLink.clickAndWait()
+
+        val propertyDetailsPage =
+            assertPageIs(
+                page,
+                PropertyDetailsPageLandlordView::class,
+                mapOf("propertyOwnershipId" to propertyOwnershipId.toString()),
+            )
+        assertThat(page.locator("main")).containsText("New invitation email sent to $invitedEmail")
+    }
+}
